Treat "alreadyExists" as an error
The create() method should only run if the resource is not in a creating
status, so it should not ever run into a cluster that was already
created by the operator. If a conflict exists it should be treated as an
error and not try to recreate or take ownership of the cluster.