Add more support for cluster updates
Without this patch, the operator only supports updating the control
plane kubernetes version and no other cluster attributes. This change
adds the ability to update the following settings:

* kubernetes version for node pools
* cluster addons
* network policy
* master authorized networks
* node pool size, autoscaling, and image type

Other attributes are immutable and are SILENTLY IGNORED if changes are
applied. This is because for many attributes if no value is set during
cluster creation then it is initialized to something upstream, so the
upstream setting is not comparable to the downstream nil or empty
setting.

The ordering matters. Some attributes need to be updated before others
(for example, control plane version must happen before node pool
version), and in some cases the API will not accept a request to update
multiple attributes at once. The update method knows to wait and requeue
further updates.