Fix ClusterAddons JSON tags
HTTPLoadBalancer was missing a doube quote in the JSON tag which was
causing the controller to treat it as a capitalized attribute in the CRD
and was making it impossible to update. Fix the syntax and also add
"omitempty" to all the cluster addons so they don't need to be rendered
if empty.