Add MonitoringService and OauthScopes attributes
MonitoringService and LoggingService are tied to each other: GKE
requires being explicit about setting both if one is changed. Add this
new attribute in order to be able to support updating both attributes in
a later commit.

Monitoring and Logging require nodes to have specific OauthScopes, so
add a parameter to handle that as well.