Make nilable attributes consistent
In GKEClusterConfigSpec, attributes that reflect syncable cluster
attributes should be pointers. This change changes the ClusterAddons
attribute to be a pointer to a struct so that it is consistent with the
other struct attributes in Spec. For all the attributes of those
sub-structs, make those NOT pointers. This makes the schema consistent
with ifself, it makes validation simpler, and makes it easier to pass
attributes directly to the GKE API since there's no need to convert
them.