Rename networkConfig to gkeClusterNetworkConfig
Without this patch, when this operator is included as a library in
Rancher, running `go generate` in Rancher does not distinguish between
the NetworkConfig struct defined here and the NetworkConfig struct
already defined for a different use in Rancher[1]. This causes the
GKEConfig schema to include the wrong NetworkConfig, and trying to
create a GKEConfig object as defined by the GKEClusterConfigSpec in this
operator will result in a GKEConfig with an empty NetworkConfig field in
Rancher.

This change renames the struct names, struct attributes and the JSON
tags for the attributes to {GKE,gke}ClusterNetworkConfig so that it
cannot collide with the NetworkConfig objects that are already in
Rancher.

[1] https://github.com/rancher/rancher/blob/6ea7600e6bddcd0483afcf86cd6142920601d4e5/pkg/client/generated/management/v3/zz_generated_network_config.go#L18