Remove docker-build step
The docker-publish step already builds the image, so it doesn't need a
build step beforehand. The docker-build step doesn't make sense to run
on pull_request events since the Dockerfile pulls from the master branch
of the remote github repository and not the locally checked-out
repository, so just remove it so that it doesn't block pull requests.