Move delete function to gke package
Move the WaitClusterRemoveExp from utils to the internal gke package,
rename it, and replace the ad hoc retry logic with the
ExponentialBackoff function from apimachinery. The backoff Factor and
Jitter are set to 0 because waiting for another operation to complete
before sending a delete request is expected and normal, and not due to a
transient network error where repeated requests would exacerbate a
server-side problem.

The delete function blocks waiting for the in-progress operation to
complete, in contrast to actions like creating or updating where a new
request will be enqueued to be acted upon later. This is because the
OnRemove handler will only retry if an error is returned, otherwise it
will remove the finalizer from the resource and the deletion will not be
retried even if Enqueue() or EnqueueAfter() is used.