Fix ClusterIpv4CidrBlock field
There is a cluster-level ClusterIpv4Cidr attribute for GKE clusters[1]
in addition to the attribute of the same name in IPAllocationPolicy[2].
Previously, this attribute existed in the CRD but was ignored during
cluster create. This commit adds the attribute to the cluster validation
and creation. It is immutable and so there is no update function added
for it.

[1] https://pkg.go.dev/google.golang.org/api/container/v1#Cluster
[2] https://pkg.go.dev/google.golang.org/api/container/v1#IPAllocationPolicy