Add ability to add and remove node pools
Without this commit, once a cluster is created, some attributes of its
node pools can be updated but none can be added or removed. This change
ensures the update action compares the list of node pools in the cluster
spec to the list of node pools known to GKE, and either adds, updates,
or removes node pools as needed. This also changes the UpdateNodePool*
functions to rely on the error response from an update request to
indicate whether a request should be retried, so that the full
update/add/delete process is streamlined and consistent.