Add controller.GetTokenSource function
The GKE handler in Rancher needs to be able to build an initial admin
kubeconfig in order to generate the first service account in the
cluster. Since gke-operator already knows how to get the credential
Secret and convert it to an OAuth2 token, it is convenient to expose
this as a public function that Rancher can use to authenticate to the
cluster.