Fix handling of region vs zone
In GKE, a cluster is created either in a Region or a Zone. The CRD
supports setting either and the controller validates that one but not
both is set. But without this patch, the operator was only respecting
the Region setting and never regarding the Zone. The examples also
incorrectly referred to an example Zone identifier in the Region
setting. This patch ensures that Zone will be used to identify the
cluster if it is used and fixes the examples to make sense.