Fix controller sync for empty version/image string
It is valid to create a GKE cluster with the kubernetes version, node
pool version, or node pool image type set to "". GKE will set defaults
on the backend. However, it causes problems during the update cycle. The
validateAndUpdate function on the controller cannot check whether "" is
semver compliant, and the Update* functions cannot set "" as a desired
value to update to. This change adds allowances for certain string
parameters to be empty. Also restructures the
UpdateMasterKubernetesVersion to be more Go-idiomatic by returning early
if possible.