Fix node pool creation validation
The error formatting in validateNodePoolCreateRequest was passing a
pointer to the node pool name attribute in to represent the cluster name
in the error message, which didn't make any sense. Fix the function to
accept the cluster name so that it can refer to it, and fix the name of
one of the fields it was checking.