Add support for AutoRepair and AutoUpgrade feature
Add a struct called Management as an attribute for NodePool which
supports toggling auto-repair and auto-upgrade for a node pool. Also
update the examples.

GKE defaults to setting these to true on the backend, so before this
patch, since we were not setting the Management pointer to an object for
a node pool, the node pool would be created with auto-repair and
auto-upgrade enabled. Now that we're explicitly setting it, it defaults
to disabled.