Rename GKE structs
Rancher's machine driver creates a dynamic schema with name "nodeconfig"
that collides with the static schema generated for the GKE operator's
NodeConfig struct. The result is that repeated calls to
/v3/schemas/nodeConfig may alternately return either the NodeConfig
schema from this operator or a different NodeConfig schema containing
the googleConfig schema for GCP cloud credentials.

This change renames the NodeConfig struct to avoid this collision, and
also renames every other struct defined for the operator in order to
prevent potential collisions in the future.