Fix locations attribute
Locations is a slice, so it is nilable, but an explicit empty slice
should not be treated the same as a nil slice. Without this change,
the JSON tag "omitempty" was making it so that creating a cluster with
`locations: []` was treated the same as not including locations and so
the cluster create would fail. This change fixes the JSON tag and also
adds the missing field to the cluster.json example.