Fix maintenanceWindow field
GKE allows not setting a maintenance window. In the SDK, this is done
with by not allocating a Cluster.MaintenancePolicy.Window object.
Without this change, we're always creating a maintenance window policy
even if the window is set to "", but "" is not a valid window and causes
a failure in the GKE API:

  googleapi: Error 400: Start_time must be in the '00:00' format, in GMT zone. Error: parsing time "" as "15:04": cannot parse "" as "15"

With this change, the maintenanceWindow field on the cluster is still
nilable but we don't try to use an invalid value if it's explicitly set
to empty.