Add support for Autopilot
GKE introduced a new mode called Autopilot[1] in which node pools no
longer need to be managed by the user. This change allows users to opt
into this mode. There are caveats regarding the flexibility of this
mode[2] so it is not necessarily suitable for all users and standard
mode should still be supported.

The cluster must have clusterAddons.horizontalPodAutoscaling and
clusterAddons.httpLoadBalancing enabled. The cluster must be created for
a region, not a zone, so zone must be unset. Any configured node pools
will be ignored, and setting node pools to an empty list should be
allowed in Rancher. The autopilot setting is immutable.

[1] https://cloud.google.com/blog/products/containers-kubernetes/introducing-gke-autopilot
[2] https://cloud.google.com/kubernetes-engine/docs/concepts/autopilot-overview