Ignore already exists error for CA Secret creation
When clusters are imported there might be a race condition where the
GKEClusterConfig Status doesn't successfully get saved on the first try
and the controller will retry on its own. Without this patch, if it does
this, it would reattempt to create the CA secret and end up failing
trying to do this. This change ensures that there is no reattempt to
create the same CA secret, which also helps ensure that if there is a
legitimate failure to update the cluster config's Status then that
failure message isn't clobbered by this unrelated error.