Fix clusterAddons examples
The minimal YAML and JSON examples should use example values that match
the upstream defaults, so change httpLoadBalancing and
horizontalPodAutoscaling to true for those samples. This will match the
default settings for a cluster created via the GKE console with no
parameters changed. This change updates the full example to set the
converse of these settings, and updates the JSON example to use the
minimalist settings for networkPolicyEnabled and networkPolicyConfig.