Remove version validation for updates
The validateUpdate method was adopted from eks-provider and mimicked its
kubernetes version validation, which ensures the provided version is
valid Semver and that the control plane and node versions are within a
constrained range of each other. This was causing a problem with GKE
versions as it was not always parsing the Semver components correctly
and would result in a confusing and incorrect error message like:

  versions for cluster [1.18.16-gke.300] and nodegroup [1.18.15-gke.1501] not compatible: all nodegroup kubernetes versionsmust be equal to or one minor version lower than the cluster kubernetes version

Rather than fix the version parsing, this change just removes this
validation. GKE does not place such strict constraints on the delta
between the control plane and node pool versions, you can even create a
cluster that is up to seven minor versions apart. The UI queries GKE for
valid versions to input, so as long as it does so there is no danger of
requesting an invalid version. This is also just an awkward place to do
this validation, since it's only validating particular attributes and
not the full update request, so if validation is needed it should be
done elsewhere.