Make networking attributes update-able
Without this change, if a mistake is made in configuring
IPAllocationPolicy or MasterAuthorizedNetworksConfig during cluster
creation, GKE can reject it with an HTTP 400 response, but there is no
way to fix the mistake either in the UI or even with a direct PUT
request. This is due to the norman 'noupdate' tags for these attributes,
which causes updates to be ignored. This change removes the noupdate
tags so that the fields can be edited while the cluster has not been
created in GKE yet. Once a cluster exists in GKE (even if it is in an
error state) it is still not possible to update these fields, and the
operator reconciler loop will ignore them.