Improve error message for create conflict
If a cluster is created by the operator in the same location with the
same name as a cluster that already exists in GKE but is not known to
Rancher, GKE will reject the request, but the cluster name cannot be
edited. This change improves the error message that will be surfaced in
the UI so that users know how to handle the problem.