Add norman annotations for *string fields
Without this patch, norman doesn't properly interpret *string types as
*strings and instead converts them to strings in the generated client
code, which prevents clients from treating these fields as nilable. To
fix this, we add an explicit type declaration in the norman tag as well
as a 'notnullable' annotation. This is counterintuitively used on fields
we expressly want to be nullable because otherwise norman interprets
''*string' as nullable and adds another * creating '**string' which is
not what we want.

Fixing the type parsing in norman would cause too many knock-on effects
in the other generated types in Rancher.