Error on nodepool name collisions
Previously, if a user tried to create two nodepools with the same name,
an update loop would commence where gke-operator would continually try
to upgrade a nodepool in GKE that had two configs in Rancher.

After this change, the collision will be detected and no update will
happen until the collision is fixed.