Retry on conflict during enqueueUpdate
It is very important for the status update to succeed here, otherwise
the update loop will be entered again unnecessarily. If this happens
very quickly, there may be an inconsistency between the upstreamSpec
state, the config.Spec state, and the actual upstream cluster state. It
is best to ensure that this status update does not compound on other
problems.