Add retries to UpdateLabels
Unlike other update actions, updating GKE cluster labels is so fast that
the cluster object does not enter a RECONCILING state, and so
when checkAndUpdate is reentered it proceeds to check all fields for
update again. However, it also lags a tiny amount and the label
fingerprint and labels on the upstreamSpec and even on the newly fetched
cluster object might be out of sync with the real cluster, and so an
attempt to update them again causes a fingerprint mismatch error. This
change adds a retry to compensate for the fact that UpdateLabels is
being called twice.