Fix setting failure message on empty resource
53dbce90 was an attempt to address the occurrence of resource conflict
errors during an UpdateStatus call to set the resource phase to
"updating". It helped reduce the frequency of this happening but it did
not fully address the issue because it was not the only place where the
resource state is updated.

The resource conflict error always occurs after an error has occurred
during the reconcile loop. Multiple status updates happen in quick
succession during a single loop, and upon entering the next loop, the
shared informer cache is not fully synced and the controller worker
fetches an out of date version of the resource object.

Without this change, when this happens, the recordError handler function
tries to continue persisting the error state to the resource object with
UpdateStatus, but since UpdateStatus returns an empty struct, there's no
object it can actually update at this point, and another error is
logged.

If this happens, just return the error right away instead of trying to
update an empty object.