Ignore label mismatch error
Reverts 3244dfa5a and instead catches the fingerprint mismatch error and
downgrades it to a debug log. The reasoning behind that commit still
applies - the upstream GKE cluster is slightly delayed in processing the
label updates - but the controller will naturally retry the update, so
we don't need to block on retrying ourselves. This way the equality
check will also be done again and so the cluster won't be updated twice.