Use norman:"pointer" for nullable fields
Opt into new "pointer" field attribute from norman that will convert any
type to a pointer in the generated client code. The generated client code
includes "omitempty" JSON tags, which means that for maps and slices,
unset and empty values are both sent to the server as unset, which
breaks the cluster create validation. With this change, there's now a
difference between a nil value and an empty allocated map or slice in the
client code. Strings can now also use this attribute instead of the
"nullablestrings" type (a85378a/2450a6e).